Add functions to module
