+++ src/modules/editTaskNote.js
@@ -1,0 +1,5 @@
+const editTaskNote = (task, newNote) => {
+    task.description = newNote || '';
+};
+  
+export default editTaskNote;
